Fix type conversions

+++ src/hotspot/share/gc/shenandoah/shenandoahMemoryManager.cpp
@@ -142,12 +142,12 @@ void ShenandoahOldGenMemoryManager::gc_end(bool recordPostGCUsage, bool recordAc
   ShenandoahMemoryManager::gc_end(recordPostGCUsage, recordAccumulatedGCTime, recordGCEndTime, countCollection, cause, allMemoryPoolsAffected);
 }
 
-size_t ShenandoahOldGenMemoryManager::ext_attribute_info_size() {
+jlong ShenandoahOldGenMemoryManager::ext_attribute_info_size() {
   return ShenandoahMemoryManager::ext_attribute_info_size() + 2;
 }
 
-size_t ShenandoahOldGenMemoryManager::ext_attribute_info(jmmExtAttributeInfo* info, jint count) {
-  size_t base = ShenandoahMemoryManager::ext_attribute_info(info, count);
+jlong ShenandoahOldGenMemoryManager::ext_attribute_info(jmmExtAttributeInfo* info, jint count) {
+  jlong base = ShenandoahMemoryManager::ext_attribute_info(info, count);
 
   info[base].name = "interrupted";
   info[base].type = 'Z';
@@ -160,8 +160,8 @@ size_t ShenandoahOldGenMemoryManager::ext_attribute_info(jmmExtAttributeInfo* in
   return base + 2;
 }
 
-size_t ShenandoahOldGenMemoryManager::ext_attribute_values(jvalue* ext_attribute_values) {
-  size_t base = ShenandoahMemoryManager::ext_attribute_values(ext_attribute_values);
+jlong ShenandoahOldGenMemoryManager::ext_attribute_values(jvalue* ext_attribute_values) {
+  jlong base = ShenandoahMemoryManager::ext_attribute_values(ext_attribute_values);
 
   ext_attribute_values[base].z = _interrupted;
   ext_attribute_values[base + 1].j = _completed_cycles;
